Tests/Day17Tests.cs
using Puzzles;

namespace Tests;

[TestFixture]
public class Day17Tests : DayTestBase<Day17>
{
    [TestCase(@"2413432311323
3215453535623
3255245654254
3446585845452
4546657867536
1438598798454
4457876987766
3637877979653
4654967986887
4564679986453
1224686865563
2546548887735
4322674655533", "102")]
    public void PuzzleOneTest(string input, string output) => Assert.That(GetPuzzle(input).GetPuzzleOneSolution(), Is.EqualTo(output));

    [TestCase(@"2413432311323
3215453535623
3255245654254
3446585845452
4546657867536
1438598798454
4457876987766
3637877979653
4654967986887
4564679986453
1224686865563
2546548887735
4322674655533", "94")]
    public void PuzzleTwoTestMix(string input, string output) => Assert.That(GetPuzzle(input).GetPuzzleTwoSolution(), Is.EqualTo(output));
    
    [TestCase(@"111111111111
999999999991
999999999991
999999999991
999999999991", "71")]
    public void PuzzleTwoTestSimple(string input, string output) => Assert.That(GetPuzzle(input).GetPuzzleTwoSolution(), Is.EqualTo(output));
}